Write module 1
